Begin Writer dev, update GeneralDirectives
-Add Settings-file to ShaderTools
-Change unit-representing types to ints from doubles
-Create interface for Texture to allow implicit textures

+++ ShaderTools/Shader/General/FogParms.cs
@@ -15,7 +15,7 @@ namespace ShaderTools.Shader.General
         /// Distance to opaque.
         /// </summary>
         [Display(Name = "Opacity", Description = "Distance (in units) to opaque.")]
-        [Range(0, double.MaxValue, ErrorMessage = "Must be positive.")]
-        public double Opacity { get; set; }
+        [Range(0, int.MaxValue, ErrorMessage = "Must be positive.")]
+        public int Opacity { get; set; }
     }
 }
